api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kidneuro"
version = "0.1.0"
description = "KidNeuro API - ASD/ADHD Therapy Edu-Games Platform"
requires-python = ">=3.12"
license = "MIT"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "gunicorn>=23.0.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "redis>=5.2.0",
    "celery[redis]>=5.4.0",
    "httpx>=0.28.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.18",
    "jinja2>=3.1.0",
    "boto3>=1.35.0",
    "sentry-sdk[fastapi]>=2.19.0",
    "structlog>=24.4.0",
    "orjson>=3.10.0",
    "typer[all]>=0.15.0",
    "rich>=13.9.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "factory-boy>=3.3.0",
    "faker>=33.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "ipython>=8.30.0",
    "pre-commit>=4.0.0",
]
prod = [
    "gunicorn>=23.0.0",
]

[project.scripts]
kidneuro = "kidneuro.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/kidneuro"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "S", "B", "A", "C4", "DTZ", "T20", "ICN", "PIE", "PT", "RSE", "RET", "SLF", "SIM", "TID", "TCH", "ARG", "PTH", "ERA"]
ignore = ["S101", "B008"]

[tool.ruff.lint.isort]
known-first-party = ["kidneuro"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
